+++ packages/shared/ReactLazyComponent.js
@@ -7,44 +7,14 @@
  * @flow
  */
 
-export type Thenable<T, R> = {
-  then(resolve: (T) => mixed, reject: (mixed) => mixed): R,
-};
+import type {
+  PendingLazyComponent,
+  ResolvedLazyComponent,
+  RejectedLazyComponent,
+  LazyComponent,
+} from 'react/src/ReactLazy';
 
-type UninitializedLazyComponent<T> = {
-  $$typeof: Symbol | number,
-  _status: -1,
-  _result: () => Thenable<{default: T, ...} | T, mixed>,
-};
-
-type PendingLazyComponent<T> = {
-  $$typeof: Symbol | number,
-  _status: 0,
-  _result: Thenable<{default: T, ...} | T, mixed>,
-};
-
-type ResolvedLazyComponent<T> = {
-  $$typeof: Symbol | number,
-  _status: 1,
-  _result: T,
-};
-
-type RejectedLazyComponent = {
-  $$typeof: Symbol | number,
-  _status: 2,
-  _result: mixed,
-};
-
-export type LazyComponent<T> =
-  | UninitializedLazyComponent<T>
-  | PendingLazyComponent<T>
-  | ResolvedLazyComponent<T>
-  | RejectedLazyComponent;
-
-export const Uninitialized = -1;
-export const Pending = 0;
-export const Resolved = 1;
-export const Rejected = 2;
+import {Uninitialized, Pending, Resolved, Rejected} from 'react/src/ReactLazy';
 
 export function refineResolvedLazyComponent<T>(
   lazyComponent: LazyComponent<T>,
@@ -59,10 +29,7 @@ export function initializeLazyComponentType(
     let ctor = lazyComponent._result;
     if (!ctor) {
       // TODO: Remove this later. THis only exists in case you use an older "react" package.
-      ctor = ((lazyComponent: any)._ctor: () => Thenable<
-        {default: any, ...} | any,
-        mixed,
-      >);
+      ctor = ((lazyComponent: any)._ctor: typeof ctor);
     }
     const thenable = ctor();
     // Transition to the next state.
